+++ api/server/router/user/user.go
@@ -41,14 +41,13 @@ func (u *userRouter) initRoutes(ginEngine *gin.Engine) {
 		// 修改密码
 		userRoute.PUT("/change/:id/password", u.changePassword)
 		// 重置密码
+		userRoute.PUT("/reset/:id/password", u.resetPassword)
 
-		//  查询当前用户角色
-		userRoute.GET("/:id/roles", u.getRolesByUserId)
-		// 根据用户id分配角色
-		userRoute.POST("/:id/roles", u.setRolesByUserId)
+		userRoute.GET("/:id/roles", u.getUserRoles)  // 查询当前用户角色
+		userRoute.POST("/:id/roles", u.setUserRoles) // 根据用户id分配角色
 
 		// 根据菜单ID获取当前用户的菜单的按钮
-		userRoute.GET("/menus/:id", u.getButtonsByCurrentUser)
+		userRoute.GET("/menus/:id/buttons", u.getButtonsByCurrentUser)
 		// 根据用户ID获取用户的菜单
 		userRoute.GET("/menus", u.getLeftMenusByCurrentUser)
 	}
