+++ docs/sql.md
@@ -27,6 +27,23 @@ CREATE TABLE `users` (
 insert into users(name, password) values ('admin', 'a66abb5684c45962d887564f08346e8d');
 ```
 
+## 创建 `roles` 表
+```sql
+CREATE TABLE `roles` (
+    id int primary key NOT NULL AUTO_INCREMENT COMMENT '主键' ,
+    gmt_create datetime COMMENT '创建时间',
+    gmt_modified datetime COMMENT '修改时间',
+    resource_version int COMMENT '版本号',
+    name varchar(128) COMMENT '用户名',
+    status tinyint(4) COMMENT '状态',
+    sequence bigint(20) NOT NULL,
+    parent_id bigint(20) NOT NULL,
+    memo varchar(128) DEFAULT NULL,
+    KEY `idx_name` (`name`),
+    UNIQUE KEY `name` (`name`)
+) ENGINE=InnoDB CHARSET=utf8 AUTO_INCREMENT=23220801;
+```
+
 ## 创建 `clouds` 表
 ```sql
 CREATE TABLE `clouds` (
