+++ pkg/db/model/user_role.go
@@ -24,6 +24,52 @@ import (
 	"github.com/caoyingjunz/gopixiu/pkg/db/gopixiu"
 )
 
+// Role 角色
+type Role struct {
+	gopixiu.Model
+
+	Memo     string `gorm:"column:memo;size:128;" json:"memo" form:"memo"`                // 备注
+	Name     string `gorm:"column:name;size:128;not null;" json:"name" form:"name"`       // 名称
+	Sequence int    `gorm:"column:sequence;not null;" json:"sequence" form:"sequence"`    // 排序值
+	ParentID int64  `gorm:"column:parent_id;not null;" json:"parent_id" form:"parent_id"` // 父级ID
+	Status   int8   `gorm:"column:status" json:"status" form:"status"`                    // 0 表示禁用，1 表示启用
+	Children []Role `gorm:"-"`
+}
+
+// TableName 自定义表名
+func (r *Role) TableName() string {
+	return "roles"
+}
+
+// BeforeCreate 创建前操作
+func (r *Role) BeforeCreate(*gorm.DB) error {
+	r.GmtCreate = time.Now()
+	r.GmtModified = time.Now()
+	return nil
+}
+
+// BeforeUpdate 更新前操作
+func (r *Role) BeforeUpdate(*gorm.DB) error {
+	r.GmtModified = time.Now()
+	return nil
+}
+
+type Rule struct {
+	gopixiu.Model
+
+	PType  string `json:"ptype" gorm:"column:ptype;size:100" description:"策略类型"`
+	Role   string `json:"role" gorm:"column:v0;size:100" description:"角色"`
+	Path   string `json:"path" gorm:"column:v1;size:100" description:"api路径"`
+	Method string `json:"method" gorm:"column:v2;size:100" description:"访问方法"`
+	V3     string `gorm:"column:v3;size:100"`
+	V4     string `gorm:"column:v4;size:100"`
+	V5     string `gorm:"column:v5;size:100"`
+}
+
+func (c *Rule) TableName() string {
+	return "rules"
+}
+
 type UserRole struct {
 	gopixiu.Model
